B – shutdown-coordinator
